Use extensionsstore for extensions management
